+++ capa-controller-role/variables.tf
@@ -36,6 +36,12 @@ variable "byovpc" {
   default     = false
 }
 
+variable "eks" {
+  type        = bool
+  description = "If true, the CAPA role will be created with the permissions needed to manage EKS clusters"
+  default     = false
+}
+
 variable "additional_policies" {
   type        = map(string)
   description = "Map of additional policy documents to attach to the IAM role"